+++ src/app/date-picker/year-picker/year-picker.component.ts
@@ -1,0 +1,84 @@
+import { Component, Input, forwardRef } from '@angular/core';
+import { CommonModule, JsonPipe } from '@angular/common';
+import {
+  DateAdapter,
+  MAT_DATE_FORMATS,
+  MAT_DATE_LOCALE,
+  MatNativeDateModule,
+} from '@angular/material/core';
+import {
+  FormControl,
+  FormsModule,
+  NG_VALUE_ACCESSOR,
+  ReactiveFormsModule,
+} from '@angular/forms';
+import { MomentDateAdapter } from '@angular/material-moment-adapter';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import {
+  MatDatepicker,
+  MatDatepickerModule,
+} from '@angular/material/datepicker';
+import moment, { Moment } from 'moment';
+import { MatInputModule } from '@angular/material/input';
+import { HeaderPickerComponent } from '../header-picker/header-picker.component';
+
+export const YEAR_MODE_FORMATS = {
+  parse: {
+    dateInput: 'YYYY',
+  },
+  display: {
+    dateInput: 'YYYY',
+    monthYearLabel: 'MMM YYYY',
+    dateA11yLabel: 'LL',
+    monthYearA11yLabel: 'MMMM YYYY',
+  },
+};
+
+@Component({
+  selector: 'app-year-picker',
+  standalone: true,
+  imports: [
+    CommonModule,
+    MatFormFieldModule,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    ReactiveFormsModule,
+    MatInputModule,
+    FormsModule,
+    JsonPipe,
+    HeaderPickerComponent,
+  ],
+  templateUrl: './year-picker.component.html',
+  styleUrl: './year-picker.component.scss',
+  providers: [
+    { provide: MAT_DATE_LOCALE, useValue: 'en' },
+    {
+      provide: DateAdapter,
+      useClass: MomentDateAdapter,
+      deps: [MAT_DATE_LOCALE],
+    },
+    { provide: MAT_DATE_FORMATS, useValue: YEAR_MODE_FORMATS },
+    {
+      provide: NG_VALUE_ACCESSOR,
+      useExisting: forwardRef(() => YearPickerComponent),
+      multi: true,
+    },
+  ],
+})
+export class YearPickerComponent {
+  customDatepickerHeader = HeaderPickerComponent;
+
+  @Input()
+  date: FormControl;
+
+  chosenYearHandler(date: Moment, dp: any) {
+    this.date.setValue(moment(date).format('yyyy'));
+    dp.close();
+  }
+
+  _openDatepickerOnClick(datepicker: MatDatepicker<Moment>) {
+    if (!datepicker.opened) {
+      datepicker.open();
+    }
+  }
+}