+++ src/app/filters/filters.component.ts
@@ -9,10 +9,10 @@ import {
 import { CommonModule } from '@angular/common';
 import {
   FormBuilder,
+  FormControl,
   FormGroup,
   FormsModule,
   ReactiveFormsModule,
-  Validators,
 } from '@angular/forms';
 import { Filter } from './filters';
 import { debounceTime } from 'rxjs';
@@ -23,6 +23,10 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { TranslateModule } from '@ngx-translate/core';
 import { setParamsMovies } from '../store/movies.actions';
+import { YearPickerComponent } from '../date-picker/year-picker/year-picker.component';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { FiltersManageComponent } from './components/filters-manage/filters-manage.component';
+import { InputRangeComponent } from './components/input-range/input-range.component';
 
 @Component({
   selector: 'app-filters',
@@ -36,6 +40,9 @@ import { setParamsMovies } from '../store/movies.actions';
     MatSelectModule,
     TranslateModule,
     MatButtonModule,
+    YearPickerComponent,
+    MatDialogModule,
+    InputRangeComponent,
   ],
   templateUrl: './filters.component.html',
   styleUrl: './filters.component.scss',
@@ -52,10 +59,10 @@ export class FiltersComponent implements OnInit {
   fb = inject(FormBuilder);
   paramsKey: any = {};
 
-  constructor() {}
+  constructor(public dialog: MatDialog) {}
 
   ngOnInit(): void {
-    this.buildFilterParams();
+    this.buildFields();
     this.form.valueChanges.pipe(debounceTime(500)).subscribe(() => {
       this.buildParams();
     });
@@ -65,12 +72,13 @@ export class FiltersComponent implements OnInit {
     let params = '';
     this.paramsKey = null;
     this.filters.forEach((filter) => {
-      if (this.form.get(filter.key)?.value) {
+      const item = this.form.get(filter.key)?.value;
+      if (item.value) {
         this.paramsKey = {
           ...this.paramsKey,
-          [filter.filterKey]: this.form.get(filter.key)?.value,
+          [filter.filterKey]: item.value,
         };
-        params = `${params}&${filter.key}=${this.form.get(filter.key)?.value}`;
+        params = `${params}&${filter.key}=${item.value}`;
       }
     });
 
@@ -81,14 +89,55 @@ export class FiltersComponent implements OnInit {
     );
   }
 
-  private buildFilterParams(): void {
+  private buildFields(): void {
     let formInputs: any = {};
+
     this.filters.forEach((filter) => {
-      formInputs[filter.key] = [''];
+      formInputs[filter.key] = new FormGroup({
+        value: new FormControl(''),
+        show: new FormControl(true),
+        data: new FormControl(filter),
+      });
       if (filter.required) {
-        formInputs[filter.key] = ['', Validators.required];
+        formInputs[filter.key] = new FormGroup({
+          value: new FormControl(null),
+          show: new FormControl(true),
+          data: new FormControl(filter),
+        });
       }
     });
     this.form = this.fb.group(formInputs);
+  }
+
+  public get year() {
+    return this.form.get('y')?.get('value') as FormControl;
+  }
+
+  public get rating(): FormGroup {
+    return this.form.get('rating') as FormGroup;
+  }
+
+  clear(): void {
+    this.filters.forEach((f) => {
+      this.form.get(f.key)?.get('value')?.setValue(null);
+    });
+
+    this.store.dispatch(
+      setParamsMovies({
+        params: '',
+      })
+    );
+    this.search.emit();
+  }
+
+  openSettingDialog(): void {
+    this.dialog.open(FiltersManageComponent, {
+      width: '550px',
+      data: {
+        form: this.form,
+        items: Object.keys(this.form.value),
+        filters: this.filters,
+      },
+    });
   }
 }