+++ src/app/store/movies.reducer.ts
@@ -1,7 +1,8 @@
 import { createReducer, on } from '@ngrx/store';
 import {
   addFavorite,
-  deleteFavorite,
+  loadFavorites,
+  loadMovies,
   setParamsMovies,
   successMovies,
   successUpdateFavorites,
@@ -9,48 +10,33 @@ import {
 } from './movies.actions';
 import { Movie } from '../shared/models/movie';
 import { Favorite } from '../shared/models/favorites';
+import { LoadingState } from './call-state';
 
 export interface MoviesState {
   params: string;
   data: Movie[];
   favorites: Favorite[];
+  callState: LoadingState;
 }
 
 export const initialState: MoviesState = {
   params: '',
   data: [],
   favorites: [],
+  callState: LoadingState.INIT,
 };
 
 export const moviesReducer = createReducer(
   initialState,
 
-  // FAVORITES
-
-  on(addFavorite, (state, { params }) => {
-    debugger;
-    return {
-      ...state,
-      favorites: [...state.favorites, params],
-    };
-  }),
-
-  on(successUpdateFavorites, (state, { data }) => {
-    return {
-      ...state,
-      data: state.data.reduce((accu: Movie[], curr: Movie) => {
-        accu.push({
-          ...curr,
-          favorite: data.find((f) => f.imdbID === curr.imdbID) ? true : false,
-        });
-        return accu;
-      }, []),
-      favorites: data,
-    };
-  }),
-
   // MOVIES
 
+  on(loadMovies, (state) => {
+    return {
+      ...state,
+      callState: LoadingState.LOADING,
+    };
+  }),
   on(setParamsMovies, (state, { params }) => {
     return {
       ...state,
@@ -61,6 +47,7 @@ export const moviesReducer = createReducer(
     return {
       ...state,
       data,
+      callState: LoadingState.LOADED,
     };
   }),
 
@@ -78,13 +65,51 @@ export const moviesReducer = createReducer(
     let favorites: Favorite[] = state.favorites;
 
     if (!state.favorites.find((f) => f.imdbID === data.imdbID)) {
-      favorites = [...state.favorites, data as Favorite];
+      favorites = [
+        ...state.favorites,
+        {
+          ...data,
+          description: '',
+          rating: 0,
+        } as Favorite,
+      ];
     }
 
     return {
       ...state,
       data: updatedItems,
       favorites,
+    };
+  }),
+
+  // FAVORITES
+
+  on(loadFavorites, (state) => {
+    return {
+      ...state,
+      callState: LoadingState.LOADING,
+    };
+  }),
+
+  on(addFavorite, (state, { params }) => {
+    return {
+      ...state,
+      favorites: [...state.favorites, params],
+    };
+  }),
+
+  on(successUpdateFavorites, (state, { data }) => {
+    return {
+      ...state,
+      data: state.data.reduce((accu: Movie[], curr: Movie) => {
+        accu.push({
+          ...curr,
+          favorite: data.find((f) => f.imdbID === curr.imdbID) ? true : false,
+        });
+        return accu;
+      }, []),
+      favorites: data,
+      callState: LoadingState.LOADED,
     };
   })
 );